Creating PR helper
